+++ tests/test_simulaton_templates.py
@@ -65,7 +65,7 @@ from threedi_modelchecker.simulation_templates.settings.extractor import (
 from threedi_modelchecker.simulation_templates.structure_controls.extractor import (
     StructureControlExtractor,
 )
-from threedi_modelchecker.threedi_model.constants import InitializationType
+from threedi_modelchecker.threedi_model.constants import InitializationType, ControlTableActionTypes
 from threedi_modelchecker.threedi_model.models import ControlGroup
 from threedi_modelchecker.threedi_model.models import ControlMeasureGroup
 from threedi_modelchecker.threedi_model.models import ControlMemory
@@ -698,7 +698,7 @@ def test_table_control_capacity_factor(session, measure_group):
         id=1, name="test group"
     )
     table_control: ControlTable = factories.ControlTableFactory.create(
-        id=1, action_table="0.0;100#1.0;200", action_type="set_capacity",
+        id=1, action_table="0.0;100#1.0;200", action_type=ControlTableActionTypes.set_capacity,
     )
 
     factories.ControlFactory.create(
@@ -718,7 +718,7 @@ def test_memory_control_capacity_factor(session, measure_group):
         id=1, name="test group"
     )
     memory_control: ControlMemory = factories.ControlMemoryFactory.create(
-        id=1, action_value="100", action_type="set_capacity",
+        id=1, action_value="100", action_type=ControlTableActionTypes.set_capacity,
     )
     factories.ControlFactory.create(
         control_group_id=control_group.id,