+++ CHANGES.rst
@@ -7,6 +7,9 @@ Changelog of threedi-modelchecker
 
 - Added checks for channel and culvert geometry distance to connection nodes.
 
+- Added checks for illegal combinations of use_2d_flow, use_1d_flow, manhole_storage_area
+  and dem file presence.
+
 
 0.20.1 (2021-12-17)
 -------------------